🔧 refactor(capa): ajustar concorrência e remover métodos não utilizados
Ajusta a concorrência do worker de 2 para 6 e remove métodos não utilizados na classe Capa.

Files Changed:
- *celery_app/main.py*: aumenta a concorrência do worker.
- *crawjud/bots/pje/capa.py*: remove método print_msg e refatora a inicialização de tarefas.
- *crawjud/common/bot.py*: remove método print_msg.
- *crawjud/core/__init__.py*: remove métodos não utilizados e simplifica importações.
- *utils/webdriver/config/chrome.py*: altera a configuração de exibição do Chrome.
