+++ utils/webdriver/config/chrome.py
@@ -15,7 +15,7 @@ work_dir = Path(__file__).cwd()
 arguments_list: list[str] = [
     "--ignore-ssl-errors=yes",
     "--ignore-certificate-errors",
-    "--display=:1",
+    "--display=:0",
     "--window-size=1600,900",
     "--no-sandbox",
     "--kiosk-printing",
